requirements.txt
streamlit
openai>=1.0.0
faiss-cpu
tqdm
requests
beautifulsoup4
youtube-transcript-api==0.6.2
pypdf